requirements.txt
osim-rl==3.0.11
scipy==1.5.4
ipykernel==5.5.6
pandas==1.1.5
tensorflow==2.4.0
tensorflow-probability==0.12.2
tqdm==4.64.0